+++ CHANGELOG.md
@@ -25,6 +25,8 @@ Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/infrastructure/compare/release-4.16.0...master
+### Fixed
+- Corrected Ubuntu codename detection
 
 
 ## [4.16.0] - 2020-06-02